scale added float

+++ scale_data.py
@@ -3,7 +3,7 @@ from imports import StandardScaler
 def scale_data(x_train, x_test, columns_to_scale):
     scaler = StandardScaler()
 
-    x_train.loc[:, columns_to_scale] = scaler.fit_transform(x_train.loc[:, columns_to_scale])
+    x_train.loc[:, columns_to_scale] = scaler.fit_transform(x_train.loc[:, columns_to_scale]).astype(float)
     x_test.loc[:, columns_to_scale] = scaler.transform(x_test.loc[:, columns_to_scale])
 
     return x_train, x_test
